capi: use c_int::from(bool) and naked signed integers for returning c_ints
c_int::from(bool) relies on the semantics documented in the Rust
reference so that:

i{32,64,etc}::from(b: bool) will return 0 with b == false and 1 with b
== true.

This change also uses `return {0,1,-1}` without an explicit From::from
for c_int as suggested by clippy. The semantics now require that a
(signed) C integer has a direct representation as a signed Rust one _as
inferred by the compiler_ (ie. i32, i64, ...).

+++ src/capi/encoding.rs
@@ -31,13 +31,13 @@ pub unsafe extern "C" fn encoding_encode_buffer(
     let c_slice_dst = if let Some(c_slice_mut) = CSliceMut::new_from_size_ptr(dst, dstlen_ptr) {
         c_slice_mut
     } else {
-        return c_int::from(-1);
+        return -1;
     };
 
     let c_slice_src = if let Some(c_slice) = CSlice::new(src, srclen) {
         c_slice
     } else {
-        return c_int::from(-1);
+        return -1;
     };
 
     let cow = encoding::encode(&c_slice_src);
@@ -46,7 +46,7 @@ pub unsafe extern "C" fn encoding_encode_buffer(
     unsafe { *dstlen_ptr = required_len };
 
     if required_len > c_slice_dst.len() {
-        return c_int::from(-1);
+        return -1;
     }
 
     unsafe {
@@ -55,5 +55,5 @@ pub unsafe extern "C" fn encoding_encode_buffer(
         core::ptr::copy_nonoverlapping(cow.as_ptr(), dst as *mut _, required_len);
     }
 
-    c_int::from(0)
+    0
 }
